refactor: centralize plainspace scripts
