+++ BlogposterCMS/public/admin.html
@@ -45,7 +45,7 @@
   <script src="/assets/js/icons.js"></script>
 
   <!-- 4) Generic page renderer (builds header, sidebar, widgets)    -->
-  <script type="module" src="/assets/js/pageRenderer.js"></script>
+  <script type="module" src="/assets/plainspace/main/pageRenderer.js"></script>
 
   <!-- Search functionality for admin top bar -->
   <script type="module" src="/assets/js/adminSearch.js"></script>
