+++ BlogposterCMS/public/index.html
@@ -19,6 +19,6 @@
   <!-- global media explorer -->
   <script src="/assets/js/openExplorer.js"></script>
   <!-- Your pageRenderer, which now just calls miniEmitter(...) -->
-  <script type="module" src="/assets/js/pageRenderer.js"></script>
+  <script type="module" src="/assets/plainspace/main/pageRenderer.js"></script>
 </body>
 </html>
